requirements.txt
altair==4.2.2
bertopic==0.14.1
joblib==1.2.0
nltk==3.8.1
numpy==1.21.6
pandas==1.2.3
Pillow==9.5.0
scikit_learn==1.0.2
sentence_transformers==2.2.2
streamlit==1.17.0
umap_learn==0.5.3
word2number==1.1
xgboost==1.4.2
